Added some additional base to handle content.

+++ android/src/com/example/mineskineditorlibgdx/model/FilesDetailsChunk.kt
@@ -6,7 +6,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class FilesDetailsChunk(
     val paths: List<String>,
-    val limit: Int = 0,
+    val limit: Long = 0,
     val cursor: String?,
     val hasMore: Boolean
 ) : Parcelable
